+++ assets/js/project-picker/index.ts
@@ -1,1 +1,2 @@
 export { ProjectPicker } from './ProjectPicker';
+export { ProjectPickerButton } from './ProjectPickerButton';