+++ lib/lightning_web/components/layouts.ex
@@ -9,7 +9,6 @@ defmodule LightningWeb.Layouts do
   def root(assigns)
 
   attr :side_menu_theme, :string, default: "primary-theme"
-  attr :theme_style, :string, default: nil
   def live(assigns)
 
   attr :side_menu_theme, :string, default: "sudo-variant"