+++ lib/lightning_web/hooks.ex
@@ -14,7 +14,6 @@ defmodule LightningWeb.Hooks do
   alias Lightning.Projects.ProjectLimiter
   alias Lightning.Services.UsageLimiter
   alias Lightning.VersionControl.VersionControlUsageLimiter
-  alias LightningWeb.Live.Helpers.ProjectTheme
   alias LightningWeb.LiveHelpers
 
   @doc """
@@ -56,20 +55,12 @@ defmodule LightningWeb.Hooks do
         {:halt, redirect(socket, to: ~p"/mfa_required")}
 
       can_access_project ->
-        scale = ProjectTheme.inline_primary_scale(project)
-
-        theme_style =
-          [scale, ProjectTheme.inline_sidebar_vars()]
-          |> Enum.reject(&is_nil/1)
-          |> Enum.join(" ")
-
         {:cont,
          socket
          |> assign(:side_menu_theme, "primary-theme")
-         |> assign(:theme_style, theme_style)
-         |> assign_new(:project_user, fn -> project_user end)
-         |> assign_new(:project, fn -> project end)
-         |> assign_new(:projects, fn -> projects end)}
+         |> assign(:project_user, project_user)
+         |> assign(:project, project)
+         |> assign(:projects, projects)}
 
       true ->
         {:halt, redirect(socket, to: "/projects") |> put_flash(:nav, :not_found)}
@@ -77,7 +68,7 @@ defmodule LightningWeb.Hooks do
   end
 
   def on_mount(:project_scope, _, _session, socket) do
-    {:cont, assign_new(socket, :theme_style, fn -> nil end)}
+    {:cont, socket}
   end
 
   def on_mount(:assign_projects, _, _session, socket) do